P001/requirements.txt
opencv-python
mediapipe
cv2-tools